+++ src/engine/m_misc.c
@@ -243,9 +243,9 @@ int M_ReadFile(int8_t const* name, byte** buffer) {
 // M_FileLength
 //
 
-intptr_t M_FileLength(FILE* handle) {
-	intptr_t savedpos;
-	intptr_t length;
+long M_FileLength(FILE* handle) {
+	long savedpos;
+	long length;
 
 	// save the current position in the file
 	savedpos = ftell(handle);
